Let BLAS handle matrix transpose
Current implementation would produce unexpected results if using a
non-BLAS function with the transpose T() matrix. `_transpose_pending` is
supposed to solve that.

I don't like having an inner matrix for the transpose, it costs
sizeof(Matrix) extra memory, when we could handle that the same matrix
itself.

In case of reusable matrixes, storing a reference to T() could cause
unexpected behaviour.